extraction of contstants
