+++ packages/packages/core/editor-app-bar/src/extensions/feedback/feedback-consts.ts
@@ -1,0 +1,2 @@
+export const EXPERIMENT_NAME = 'in_editor_feedback';
+export const FEEDBACK_TOGGLE_EVENT = 'elementor/open-feedback';
