+++ packages/packages/core/editor-app-bar/src/extensions/feedback/index.ts
@@ -2,10 +2,10 @@ import { isExperimentActive } from '@elementor/editor-v1-adapters';
 import { MessageLinesIcon } from '@elementor/icons';
 import { __ } from '@wordpress/i18n';
 
-import { EXPERIMENT_NAME } from '../../components/locations/send-feedback-popup-location';
 import { mainMenu } from '../../locations';
+import { EXPERIMENT_NAME, FEEDBACK_TOGGLE_EVENT } from './feedback-consts';
 
-export const FEEDBACK_TOGGLE_EVENT = 'elementor/open-feedback';
+
 
 export function init() {
 	const isActive = isExperimentActive( EXPERIMENT_NAME );
